Migrated to Maven for builds.
Refactored package name to match domain.
Added .gitattributes and .editorconfig.
Updated .gitignore for Intellij and Maven.
