+++ plugin/.gitignore
@@ -1,4 +1,20 @@
-/bin
-/out
-/upload.sh
-/utils
+# Maven
+target/
+dependency-reduced-pom.xml
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+release.properties
+
+# IntelliJ IDEA
+.idea/
+*.iml
+*.ipr
+*.iws
+
+# Eclipse
+.settings/
+.classpath
+.project
+
+# Other
+upload.sh
