+++ plugin/src/main/resources/plugin.yml
@@ -1,0 +1,36 @@
+name: EpicStructureManagement
+main: com.playmonumenta.epicstructures.Plugin
+version: 1.0
+commands:
+  loadStructure:
+    permission: epicstructures
+    description: Loads the specified structure at the indicated coordinates
+    usage: /loadStructure <name> <x> <y> <z>
+  saveStructure:
+    permission: epicstructures
+    description: Saves the specified structure contained by the given coordinates
+    usage: /saveStructure <name> <x1> <y1> <z1> <x2> <y2> <z2>
+  respawnStructure:
+    permission: epicstructures
+    description: Sets the countdown timer on the specified structure
+    usage: /respawnStructure <name> [ticks=600]
+  addRespawningStructure:
+    permission: epicstructures
+    description: Adds a new automatically respawning structure
+    usage: /addRespawningStructure <label> <path> <x> <y> <z> <extraRadius> <respawnTime> <display name with spaces>
+  removeRespawningStructure:
+    permission: epicstructures
+    description: Removes an automatically respawning structure
+    usage: /removeRespawningStructure <label>
+  listRespawningStructures:
+    permission: epicstructures
+    description: Lists automatically respawning structures
+    usage: /listRespawningStructures
+  setPostRespawnCommand:
+    permission: epicstructures
+    description: Sets the command run after respawning a structure
+    usage: /setPostRespawnCommand <label> [command with spaces]
+  activateSpecialStructure:
+    permission: epicstructures
+    description: Sets the path of the next respawn of the structure.
+    usage: /activateSpecialStructure <label> [path_to_special_structure]
